+++ packages/xbell/src/types/expect.ts
@@ -11,7 +11,7 @@ export interface E2EMatcher extends ExpectMatchObject {
   toBeDisabled: (received: Locator | ElementHandle, options?: TimeoutOptions) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
   toBeVisible: (received: Locator | ElementHandle, options?: TimeoutOptions) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
   toBeHidden: (received: Locator | ElementHandle, options?: TimeoutOptions) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
-  toMatchImageScreenshot: (received: Uint8Array | Buffer, options: ToMatchImageSnapshotOptions | string) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
+  toMatchImageSnapshot: (received: Uint8Array | Buffer, options: ToMatchImageSnapshotOptions | string) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
   toMatchSnapshot: (received: any, options: ToMatchJavaScriptSnapshotOptions | string) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
   toThrowErrorMatchingSnapshot: (received: Function | Error, options: ToMatchJavaScriptSnapshotOptions) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;
   toMatchScreenshot: (received: Locator | ElementHandle | CommonPage, options: ToMatchImageSnapshotOptions | string) => ReturnType<ExpectMatchFunctionReturnPromiseFunction>;