+++ packages/xbell/src/types/frame-locator.ts
@@ -5,6 +5,7 @@ export interface FrameLocator {
   getByTestId(testId: string): Locator;
   getByClass(className: string): Locator;
   get(selector: string): Locator;
+  locator(selector: string): Locator;
   getById(id: string): Locator;
   first(): FrameLocator;
   last(): FrameLocator;