+++ tests/browser-canvas-to-image/xbell.config.ts
@@ -1,0 +1,9 @@
+import type { XBellConfig } from 'xbell';
+
+const config: XBellConfig = {
+  browser: {
+    headless: false,
+  }
+};
+
+export default config;