+++ tests/browser-debug/src/main.test.ts
@@ -1,0 +1,7 @@
+import { test } from 'xbell';
+
+test.browser('browser debug', () => {
+  const a = 1 + 1;
+  debugger;
+  console.log(a);
+});